test: fix autoloaded classes
Deprecation Notice: Class `ClassName` located in `class.php` does not comply with psr-4 autoloading standard. It will not autoload anymore in Composer v2.0.
