+++ tests/_data/fixtures/Traits/Db/MysqlTrait.php
@@ -12,11 +12,14 @@ declare(strict_types=1);
 
 namespace Phalcon\Test\Fixtures\Traits\Db;
 
-use Phalcon\Db\Column;
 use Phalcon\Db\Index;
 use Phalcon\Db\Reference;
-use function array_shift;
 
+/**
+ * Trait MysqlTrait
+ *
+ * @package Phalcon\Test\Fixtures\Traits\Db
+ */
 trait MysqlTrait
 {
     protected $connection = null;
