--- tests/_data/fixtures/Traits/_tocheck/CryptTrait.php
@@ -1,23 +1,0 @@
-<?php
-/**
- * This file is part of the Phalcon Framework.
- *
- * (c) Phalcon Team <team@phalcon.io>
- *
- * For the full copyright and license information, please view the LICENSE.txt
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Phalcon\Test\Fixtures\Traits;
-
-use UnitTester;
-
-trait CryptTrait
-{
-    public function _before(UnitTester $I)
-    {
-        $I->checkExtensionIsLoaded('openssl');
-    }
-}
