+++ tests/integration/Image/ImageFactory/LoadCest.php
