+++ tests/integration/Mvc/Dispatcher/Refactor/DispatcherAfterDispatchCest.php
@@ -1,53 +1,50 @@
 <?php
 
-namespace Phalcon\Test\Integration\Mvc\Dispatcher;
+/**
+ * This file is part of the Phalcon Framework.
+ *
+ * (c) Phalcon Team <team@phalcon.io>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Phalcon\Test\Integration\Mvc\Dispatcher\Refactor;
 
 use Exception;
 use IntegrationTester;
-use Phalcon\Test\Integration\Mvc\Dispatcher\Helper\BaseDispatcher;
+use Phalcon\Test\Integration\Mvc\Dispatcher\Helper\BaseDispatcher;;
 
 /**
- * \Phalcon\Test\Integration\Mvc\Dispatcher\DispatcherAfterExecuteRouteTest
- * Tests the \Phalcon\Dispatcher and Phalcon\Mvc\Dispatcher "afterExecuteRoute"
- * event.
+ * Class DispatcherAfterDispatchCest
  *
- * @link          https://docs.phalcon.io/en/latest/reference/dispatching.html
- *
- * @copyright (c) 2011-2017 Phalcon Team
- * @link          http://www.phalcon.io
- * @author        Andres Gutierrez <andres@phalcon.io>
- * @author        Nikolaos Dimopoulos <nikos@phalcon.io>
- *
- * The contents of this file are subject to the New BSD License that is
- * bundled with this package in the file docs/LICENSE.txt
- *
- * If you did not receive a copy of the license and are unable to obtain it
- * through the world-wide-web, please send an email to license@phalcon.io
- * so that we can send you a copy immediately.
+ * @package Phalcon\Test\Integration\Mvc\Dispatcher
+ * @todo: refactor
  */
-class DispatcherAfterExecuteRouteCest extends BaseDispatcher
+class DispatcherAfterDispatchCest extends BaseDispatcher
 {
     /**
-     * Tests the forwarding in the afterExecuteRoute event
+     * Tests the forwarding in the afterDispatch event
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterExecuteRouteForwardOnce(IntegrationTester $I)
+    public function testAfterDispatchForwardOnce(IntegrationTester $I)
     {
         $forwarded  = false;
         $dispatcher = $this->getDispatcher();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterExecuteRoute',
+            'dispatch:afterDispatch',
             function ($event, $dispatcher) use (&$forwarded) {
                 if ($forwarded === false) {
                     $dispatcher->forward(['action' => 'index2']);
                     $forwarded = true;
                 }
             }
-        )
-        ;
+        );
 
         $dispatcher->dispatch();
 
@@ -60,6 +57,8 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
             'afterInitialize',
             'indexAction',
             'afterExecuteRoute',
+            'afterExecuteRoute-method',
+            'afterDispatch',
             'beforeDispatch',
             'beforeExecuteRoute',
             'beforeExecuteRoute-method',
@@ -69,34 +68,29 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
             'afterDispatch',
             'afterDispatchLoop',
         ];
-
-        $I->assertEquals(
-            $expected,
-            $this->getDispatcherListener()->getTrace()
-        );
+        $actual   = $this->getDispatcherListener()->getTrace();
+        $I->assertEquals($expected, $actual);
     }
 
     /**
-     * Tests returning <tt>false</tt> inside a afterExecuteRoute event.
+     * Tests returning <tt>false</tt> inside a afterDispatch event.
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterExecuteRouteReturnFalse(IntegrationTester $I)
+    public function testAfterDispatchReturnFalse(IntegrationTester $I)
     {
         $dispatcher         = $this->getDispatcher();
         $dispatcherListener = $this->getDispatcherListener();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterExecuteRoute',
+            'dispatch:afterDispatch',
             function () use ($dispatcherListener) {
-                $dispatcherListener->trace('afterExecuteRoute: custom return false');
+                $dispatcherListener->trace('afterDispatch: custom return false');
 
                 return false;
             }
-        )
-        ;
-
+        );
         $dispatcher->dispatch();
 
         $expected = [
@@ -108,43 +102,40 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
             'afterInitialize',
             'indexAction',
             'afterExecuteRoute',
-            'afterExecuteRoute: custom return false',
+            'afterExecuteRoute-method',
+            'afterDispatch',
+            'afterDispatch: custom return false',
             'afterDispatchLoop',
         ];
-
-        $I->assertEquals(
-            $expected,
-            $this->getDispatcherListener()->getTrace()
-        );
+        $actual   = $this->getDispatcherListener()->getTrace();
+        $I->assertEquals($expected, $actual);
     }
 
     /**
      * Tests exception handling to ensure exceptions can be properly handled
-     * when thrown from inside a afterExecuteRoute event and then ensure the
+     * when thrown from inside an "afterDispatch" event and then ensure the
      * exception is not bubbled when returning with <tt>false</tt>.
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterExecuteRouteWithBeforeExceptionReturningFalse(IntegrationTester $I)
+    public function testAfterDispatchWithBeforeExceptionReturningFalse(IntegrationTester $I)
     {
         $dispatcher = $this->getDispatcher();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterExecuteRoute',
+            'dispatch:afterDispatch',
             function () {
-                throw new Exception('afterExecuteRoute exception occurred');
+                throw new Exception('afterDispatch exception occurred');
             }
-        )
-        ;
+        );
 
         $dispatcher->getEventsManager()->attach(
             'dispatch:beforeException',
             function () {
                 return false;
             }
-        )
-        ;
+        );
 
         $dispatcher->dispatch();
 
@@ -157,14 +148,13 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
             'afterInitialize',
             'indexAction',
             'afterExecuteRoute',
-            'beforeException: afterExecuteRoute exception occurred',
+            'afterExecuteRoute-method',
+            'afterDispatch',
+            'beforeException: afterDispatch exception occurred',
             'afterDispatchLoop',
         ];
-
-        $I->assertEquals(
-            $expected,
-            $this->getDispatcherListener()->getTrace()
-        );
+        $actual   = $this->getDispatcherListener()->getTrace();
+        $I->assertEquals($expected, $actual);
     }
 
     /**
@@ -175,19 +165,17 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterExecuteRouteWithBeforeExceptionBubble(IntegrationTester $I)
+    public function testAfterDispatchWithBeforeExceptionBubble(IntegrationTester $I)
     {
         $dispatcher         = $this->getDispatcher();
         $dispatcherListener = $this->getDispatcherListener();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterExecuteRoute',
+            'dispatch:afterDispatch',
             function () {
-                throw new Exception('afterExecuteRoute exception occurred');
+                throw new Exception('afterDispatch exception occurred');
             }
-        )
-        ;
-
+        );
         $dispatcher->getEventsManager()->attach(
             'dispatch:beforeException',
             function () use ($dispatcherListener) {
@@ -195,8 +183,7 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
 
                 return null;
             }
-        )
-        ;
+        );
 
         $I->expectThrowable(
             Exception::class,
@@ -214,54 +201,46 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
             'afterInitialize',
             'indexAction',
             'afterExecuteRoute',
-            'beforeException: afterExecuteRoute exception occurred',
+            'afterExecuteRoute-method',
+            'afterDispatch',
+            'beforeException: afterDispatch exception occurred',
             'beforeException: custom before exception bubble',
         ];
-
-        $I->assertEquals(
-            $expected,
-            $this->getDispatcherListener()->getTrace()
-        );
+        $actual   = $this->getDispatcherListener()->getTrace();
+        $I->assertEquals($expected, $actual);
     }
 
     /**
      * Tests dispatch forward handling inside the beforeException when a
-     * afterExecuteRoute exception occurs.
+     * afterDispatch exception occurs.
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterExecuteRouteWithBeforeExceptionForwardOnce(IntegrationTester $I)
+    public function testAfterDispatchWithBeforeExceptionForwardOnce(IntegrationTester $I)
     {
         $forwarded          = false;
         $dispatcher         = $this->getDispatcher();
         $dispatcherListener = $this->getDispatcherListener();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterExecuteRoute',
+            'dispatch:afterDispatch',
             function () use (&$forwarded) {
                 if ($forwarded === false) {
                     $forwarded = true;
 
-                    throw new Exception('afterExecuteRoute exception occurred');
+                    throw new Exception('afterDispatch exception occurred');
                 }
             }
-        )
-        ;
+        );
 
         $dispatcher->getEventsManager()->attach(
             'dispatch:beforeException',
             function ($event, $dispatcher) use ($dispatcherListener) {
                 $dispatcherListener->trace('beforeException: custom before exception forward');
-
-                $dispatcher->forward(
-                    [
-                        'action' => 'index2',
-                    ]
-                );
+                $dispatcher->forward(['action' => 'index2']);
             }
-        )
-        ;
+        );
 
         $dispatcher->dispatch();
 
@@ -274,7 +253,9 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
             'afterInitialize',
             'indexAction',
             'afterExecuteRoute',
-            'beforeException: afterExecuteRoute exception occurred',
+            'afterExecuteRoute-method',
+            'afterDispatch',
+            'beforeException: afterDispatch exception occurred',
             'beforeException: custom before exception forward',
             'beforeDispatch',
             'beforeExecuteRoute',
@@ -285,10 +266,7 @@ class DispatcherAfterExecuteRouteCest extends BaseDispatcher
             'afterDispatch',
             'afterDispatchLoop',
         ];
-
-        $I->assertEquals(
-            $expected,
-            $this->getDispatcherListener()->getTrace()
-        );
+        $actual   = $this->getDispatcherListener()->getTrace();
+        $I->assertEquals($expected, $actual);
     }
 }
