+++ tests/integration/Mvc/Dispatcher/Refactor/DispatcherAfterDispatchLoopCest.php
@@ -1,45 +1,43 @@
 <?php
 
-namespace Phalcon\Test\Integration\Mvc\Dispatcher;
+/**
+ * This file is part of the Phalcon Framework.
+ *
+ * (c) Phalcon Team <team@phalcon.io>
+ *
+ * For the full copyright and license information, please view the LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Phalcon\Test\Integration\Mvc\Dispatcher\Refactor;
 
 use Exception;
 use IntegrationTester;
-use Phalcon\Test\Integration\Mvc\Dispatcher\Helper\BaseDispatcher;
+use Phalcon\Test\Integration\Mvc\Dispatcher\Helper\BaseDispatcher;;
 
 /**
- * \Phalcon\Test\Integration\Mvc\Dispatcher\DispatcherAfterInitializeTest
- * Tests the \Phalcon\Dispatcher and Phalcon\Mvc\Dispatcher "afterInitialize"
- * event.
+ * Class DispatcherAfterDispatchLoopCest
  *
- * @link          https://docs.phalcon.io/en/latest/reference/dispatching.html
- *
- * @copyright (c) 2011-2017 Phalcon Team
- * @link          http://www.phalcon.io
- * @author        Andres Gutierrez <andres@phalcon.io>
- * @author        Nikolaos Dimopoulos <nikos@phalcon.io>
- *
- * The contents of this file are subject to the New BSD License that is
- * bundled with this package in the file docs/LICENSE.txt
- *
- * If you did not receive a copy of the license and are unable to obtain it
- * through the world-wide-web, please send an email to license@phalcon.io
- * so that we can send you a copy immediately.
+ * @package Phalcon\Test\Integration\Mvc\Dispatcher
+ * @todo: refactor
  */
-class DispatcherAfterInitializeCest extends BaseDispatcher
+class DispatcherAfterDispatchLoopCest extends BaseDispatcher
 {
     /**
-     * Tests the forwarding in the afterInitialize event
+     * Tests the forwarding in the afterDispatchLoop event
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterInitializeForwardOnce(IntegrationTester $I)
+    public function testAfterDispatchLoopForward(IntegrationTester $I)
     {
         $forwarded  = false;
         $dispatcher = $this->getDispatcher();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterInitialize',
+            'dispatch:afterDispatchLoop',
             function ($event, $dispatcher) use (&$forwarded) {
                 if ($forwarded === false) {
                     $dispatcher->forward(
@@ -51,8 +49,7 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
                     $forwarded = true;
                 }
             }
-        )
-        ;
+        );
 
         $dispatcher->dispatch();
 
@@ -63,10 +60,7 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
             'beforeExecuteRoute-method',
             'initialize-method',
             'afterInitialize',
-            'beforeDispatch',
-            'beforeExecuteRoute',
-            'beforeExecuteRoute-method',
-            'index2Action',
+            'indexAction',
             'afterExecuteRoute',
             'afterExecuteRoute-method',
             'afterDispatch',
@@ -80,25 +74,24 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
     }
 
     /**
-     * Tests returning <tt>false</tt> inside a afterInitialize event.
+     * Tests returning <tt>false</tt> inside an afterDispatchLoop event.
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterInitializeReturnFalse(IntegrationTester $I)
+    public function testAfterDispatchLoopReturnFalse(IntegrationTester $I)
     {
         $dispatcher         = $this->getDispatcher();
         $dispatcherListener = $this->getDispatcherListener();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterInitialize',
+            'dispatch:afterDispatchLoop',
             function () use ($dispatcherListener) {
-                $dispatcherListener->trace('afterInitialize: custom return false');
+                $dispatcherListener->trace('afterDispatchLoop: custom return false');
 
                 return false;
             }
-        )
-        ;
+        );
 
         $dispatcher->dispatch();
 
@@ -109,8 +102,12 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
             'beforeExecuteRoute-method',
             'initialize-method',
             'afterInitialize',
-            'afterInitialize: custom return false',
+            'indexAction',
+            'afterExecuteRoute',
+            'afterExecuteRoute-method',
+            'afterDispatch',
             'afterDispatchLoop',
+            'afterDispatchLoop: custom return false',
         ];
 
         $I->assertEquals(
@@ -121,31 +118,29 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
 
     /**
      * Tests exception handling to ensure exceptions can be properly handled
-     * when thrown from inside a afterInitialize event and then ensure the
+     * when thrown from inside an "afterDispatchLoop" event and then ensure the
      * exception is not bubbled when returning with <tt>false</tt>.
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterInitializeWithBeforeExceptionReturningFalse(IntegrationTester $I)
+    public function testAfterDispatchLoopWithBeforeExceptionReturningFalse(IntegrationTester $I)
     {
         $dispatcher = $this->getDispatcher();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterInitialize',
+            'dispatch:afterDispatchLoop',
             function () {
-                throw new Exception('afterInitialize exception occurred');
+                throw new Exception('afterDispatch exception occurred');
             }
-        )
-        ;
+        );
 
         $dispatcher->getEventsManager()->attach(
             'dispatch:beforeException',
             function () {
                 return false;
             }
-        )
-        ;
+        );
 
         $dispatcher->dispatch();
 
@@ -156,8 +151,12 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
             'beforeExecuteRoute-method',
             'initialize-method',
             'afterInitialize',
-            'beforeException: afterInitialize exception occurred',
+            'indexAction',
+            'afterExecuteRoute',
+            'afterExecuteRoute-method',
+            'afterDispatch',
             'afterDispatchLoop',
+            'beforeException: afterDispatch exception occurred',
         ];
 
         $I->assertEquals(
@@ -174,19 +173,17 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterInitializeWithBeforeExceptionBubble(IntegrationTester $I)
+    public function testAfterDispatchLoopWithBeforeExceptionBubble(IntegrationTester $I)
     {
         $dispatcher         = $this->getDispatcher();
         $dispatcherListener = $this->getDispatcherListener();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterInitialize',
+            'dispatch:afterDispatchLoop',
             function () {
-                throw new Exception('afterInitialize exception occurred');
+                throw new Exception('afterDispatchLoop exception occurred');
             }
-        )
-        ;
-
+        );
         $dispatcher->getEventsManager()->attach(
             'dispatch:beforeException',
             function () use ($dispatcherListener) {
@@ -196,8 +193,7 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
 
                 return null;
             }
-        )
-        ;
+        );
 
         $I->expectThrowable(
             Exception::class,
@@ -213,7 +209,12 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
             'beforeExecuteRoute-method',
             'initialize-method',
             'afterInitialize',
-            'beforeException: afterInitialize exception occurred',
+            'indexAction',
+            'afterExecuteRoute',
+            'afterExecuteRoute-method',
+            'afterDispatch',
+            'afterDispatchLoop',
+            'beforeException: afterDispatchLoop exception occurred',
             'beforeException: custom before exception bubble',
         ];
 
@@ -224,36 +225,33 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
     }
 
     /**
-     * Tests dispatch forward handling inside the beforeException when a
-     * afterInitialize exception occurs.
+     * Tests dispatch forward handling inside the beforeException when an
+     * exception occurs in an "afterDispatchLoop" event.
      *
      * @author Mark Johnson <https://github.com/virgofx>
      * @since  2017-10-07
      */
-    public function testAfterInitializeWithBeforeExceptionForwardOnce(IntegrationTester $I)
+    public function testAfterDispatchLoopWithBeforeExceptionForwardOnce(IntegrationTester $I)
     {
         $forwarded          = false;
         $dispatcher         = $this->getDispatcher();
         $dispatcherListener = $this->getDispatcherListener();
 
         $dispatcher->getEventsManager()->attach(
-            'dispatch:afterInitialize',
+            'dispatch:afterDispatchLoop',
             function () use (&$forwarded) {
                 if ($forwarded === false) {
                     $forwarded = true;
 
-                    throw new Exception('afterInitialize exception occurred');
+                    throw new Exception('afterDispatchLoop exception occurred');
                 }
             }
-        )
-        ;
+        );
 
         $dispatcher->getEventsManager()->attach(
             'dispatch:beforeException',
             function ($event, $dispatcher) use ($dispatcherListener) {
-                $dispatcherListener->trace(
-                    'beforeException: custom before exception forward'
-                );
+                $dispatcherListener->trace('beforeException: custom before exception forward');
 
                 $dispatcher->forward(
                     [
@@ -261,10 +259,14 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
                     ]
                 );
             }
-        )
-        ;
+        );
 
-        $dispatcher->dispatch();
+        $I->expectThrowable(
+            Exception::class,
+            function () use ($dispatcher) {
+                $dispatcher->dispatch();
+            }
+        );
 
         $expected = [
             'beforeDispatchLoop',
@@ -273,16 +275,13 @@ class DispatcherAfterInitializeCest extends BaseDispatcher
             'beforeExecuteRoute-method',
             'initialize-method',
             'afterInitialize',
-            'beforeException: afterInitialize exception occurred',
-            'beforeException: custom before exception forward',
-            'beforeDispatch',
-            'beforeExecuteRoute',
-            'beforeExecuteRoute-method',
-            'index2Action',
+            'indexAction',
             'afterExecuteRoute',
             'afterExecuteRoute-method',
             'afterDispatch',
             'afterDispatchLoop',
+            'beforeException: afterDispatchLoop exception occurred',
+            'beforeException: custom before exception forward',
         ];
 
         $I->assertEquals(
