+++ tests/integration/Storage/Serializer/Php/UnserializeCest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Phalcon\Test\Integration\Storage\Unserializer\Php;
+namespace Phalcon\Test\Integration\Storage\Serializer\Php;
 
 use Codeception\Example;
 use InvalidArgumentException;
