+++ tests/integration/Validation/Validator/StringLength/GetOptionCest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Phalcon\Test\Integration\Validation\Validator\StringLength2;
+namespace Phalcon\Test\Integration\Validation\Validator\StringLength;
 
 use IntegrationTester;
 use Phalcon\Test\Fixtures\Traits\ValidationTrait;
