+++ tests/unit/Collection/ReadOnly/ClearCest.php
