+++ tests/unit/Collection/ReadOnly/ConstructCest.php
