+++ tests/unit/Collection/ReadOnly/CountCest.php
