+++ tests/unit/Collection/ReadOnly/GetIteratorCest.php
