+++ tests/unit/Collection/ReadOnly/HasCest.php
