+++ tests/unit/Collection/ReadOnly/RemoveCest.php
