+++ tests/unit/Collection/ReadOnly/SerializeCest.php
