+++ tests/unit/Collection/ReadOnly/ToArrayCest.php
