+++ tests/unit/Collection/ReadOnly/ToJsonCest.php
