+++ tests/unit/Filter/Filter/GetSetHasCest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Phalcon\Test\Unit\Filter;
+namespace Phalcon\Test\Unit\Filter\Filter;
 
 use Closure;
 use Phalcon\Filter;
