+++ tests/unit/Http/Request/RequestCest.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace Phalcon\Test\Unit\Http;
+namespace Phalcon\Test\Unit\Http\Request;
 
 use Phalcon\Test\Unit\Http\Helper\HttpBase;
 use UnitTester;
