+++ tests/unit/Loader/LoaderCest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Phalcon\Test\Unit;
+namespace Phalcon\Test\Unit\Loader;
 
 use LoaderEvent;
 use Phalcon\Events\Event;
